ENH: API modification for DIO_Configure, allows Mask object to be passed in.

+++ AIOUSB/samples/USB-AI16-16/cust_sample.cpp
@@ -21,19 +21,6 @@
 #include <mach/clock.h>
 #include <mach/mach.h>
 #endif
-// struct timespec ts;
-
-// #ifdef __MACH__ // OS X does not have clock_gettime, use clock_get_time
-// clock_serv_t cclock;
-// mach_timespec_t mts;
-// host_get_clock_service( mach_host_self(), CALENDAR_CLOCK, &cclock );
-// clock_get_time(cclock, &mts);
-// mach_port_deallocate(mach_task_self(), cclock);
-// tv.tv_sec = mts.tv_sec;
-// tv.tv_nsec = mts.tv_nsec;
-// #else
-// clock_gettime(CLOCK_REALTIME, &tv);
-// #endif
 
 #ifdef __MACH__
 #include <sys/time.h>
@@ -705,7 +692,7 @@ void AioUsbPrintDeviceProps(DeviceProperties *devProps ) {
   fprintf(stderr, "---------- AIO USB Device Properties ----------\n");
   fprintf(stderr, "Name: %s\n", devProps->Name);
   fprintf(stderr, "ProductID: 0x%x\n", devProps->ProductID);
-  fprintf(stderr, "SerialNumber: %llu\n", devProps->SerialNumber);
+  fprintf(stderr, "SerialNumber: %d\n", (int)devProps->SerialNumber);
   fprintf(stderr, "ADCChannels: %u\n", devProps->ADCChannels);
   fprintf(stderr, "ADCChannelsPerGroup: %u\n", devProps->ADCChannelsPerGroup);
   fprintf(stderr, "ADCMUXChannels: %u\n", devProps->ADCMUXChannels);
